+++ ms-user/src/main/java/com/example/ms_user/repositories/UserRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity,Integer> {
+public interface UserRepository extends JpaRepository<UserEntity,Long> {
     @Override
-    Optional<UserEntity> findById(Integer userId);
+    Optional<UserEntity> findById(Long userId);
 }
